LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/worker-importscripts.sub-expected.txt


Harness Error (FAIL), message = Unhandled rejection: null is not an object (evaluating 'obj.addEventListener')

FAIL Dedicated worker inherits CSP from main document assert_equals: expected "importScripts allowed" but got "importScripts blocked"
FAIL Shared worker delivers its own CSP null is not an object (evaluating 'w.port.start')
